@@ -7,16 +7,16 @@
     "tools/amp-validation"
   ],
   "engines": {
-    "yarn": "1.3.2"
+    "yarn": "1.5.1"
   },
   "dependencies": {
     "@guardian/atom-renderer": "0.12.2",
     "bean": "~1.0.14",
     "bonzo": "~2.0.0",
-    "circular-dependency-plugin": "3.0.0",
     "classnames": "~2.2.0",
     "core-js": "^2.5.3",
     "curl": "cujojs/curl#0.8.9",
+    "debounce-promise": "^3.1.0",
     "domready": "^1.0.8",
     "fastdom": "0.8.5",
     "fence": "guardian/fence#0.2.11",
@@ -27,7 +27,8 @@
     "qwery": "3.4.2",
     "rangefix": "^0.2.5",
     "raven-js": "^3.19.1",
-    "react": "0.13.2",
+    "react": "^0.14.9",
+    "react-dom": "^0.14.9",
     "reqwest": "2.0.5",
     "tcp-ping": "^0.1.1",
     "video.js": "~5.3",
@@ -56,7 +57,7 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.0",
-    "babel-preset-flow": "^6.23.0",
+    "babel-preset-react": "^6.24.1",
     "babel-register": "^6.26.0",
     "browser-sync": "^2.18.13",
     "bs-fullscreen-message": "^1.1.0",
@@ -64,6 +65,7 @@
     "chalk": "^2.1.0",
     "chance": "^1.0.11",
     "chokidar": "^1.7.0",
+    "circular-dependency-plugin": "^5.0.1",
     "cp-file": "^4.2.0",
     "cpy": "^5.1.0",
     "cssstats": "^3.1.0",
@@ -117,7 +119,7 @@
     "uglify-js": "^2.7.4",
     "uglifyjs-webpack-plugin": "^1.0.0-rc.0",
     "webpack": "^4.1.1",
-    "webpack-bundle-analyzer": "^2.9.0",
+    "webpack-bundle-analyzer": "^2.11.1",
     "webpack-merge": "^4.1.0",
     "webpack-visualizer-plugin": "^0.1.11",
     "yargs": "^11.0.0"
