@@ -1,5 +1,6 @@
 // @flow
-import React from 'react/addons';
+import React, { Component } from 'react';
+import { render } from 'react-dom';
 import { saveState, isOn } from 'common/modules/accessibility/main';
 
 const DOM_ID: string = 'js-accessibility-preferences';
@@ -8,50 +9,46 @@ type AccessibilityState = {
     'flashing-elements': boolean,
 };
 
-const BinaryToggle = React.createClass({
+class BinaryToggle extends Component<*, *> {
     render() {
-        return React.DOM.div(
-            {
-                className: 'form-field',
-            },
-            React.DOM.div(
-                {
-                    className: 'checkbox',
-                },
-                React.DOM.label(
-                    {
-                        className: 'label',
-                    },
-                    [
-                        React.DOM.input({
-                            key: this.props.name,
-                            type: 'checkbox',
-                            'data-link-name': this.props.name,
-                            defaultChecked: this.props.enabled,
-                            onChange: this.props.handleChange,
+        return (
+            <div className="form-field">
+                <div className="checkbox">
+                    <label
+                        className="label"
+                        htmlFor={`checkbox-${this.props.name}`}>
+                        <input
+                            id={`checkbox-${this.props.name}`}
+                            key={this.props.name}
+                            type="checkbox"
+                            data-link-name={this.props.name}
+                            defaultChecked={this.props.enabled}
+                            onChange={this.props.handleChange.bind(this)}
                             // TODO damn me when I decided to implement this page
                             // global.css includes pasteup-forms 5 which applies ugly styles
                             // on all inputs. The proper solution is to upgrade pasteup to 6
                             // but I prefer to leave the pleasure to someone else
                             // The style object should be removed once we upgrade. Never probably.
-                            style: {
+                            style={{
                                 float: 'none',
                                 margin: '3px 0.5ex',
-                            },
-                        }),
-                    ].concat(this.props.label)
-                )
-            )
+                            }}
+                        />
+                        {this.props.label}
+                    </label>
+                </div>
+            </div>
         );
-    },
-});
+    }
+}
 
-const Accessibility = React.createClass({
-    getInitialState(): AccessibilityState {
-        return {
+class Accessibility extends Component<*, *> {
+    constructor() {
+        super();
+        this.state = ({
             'flashing-elements': isOn('flashing-elements'),
-        };
-    },
+        }: AccessibilityState);
+    }
 
     toggle(key: string): void {
         const newState = {};
@@ -59,62 +56,52 @@ const Accessibility = React.createClass({
         this.setState(newState, function() {
             saveState(this.state);
         });
-    },
+    }
 
     render(): Object {
-        return React.DOM.form(
-            {
-                className: 'form',
-            },
-            React.DOM.fieldset(
-                {
-                    className: 'fieldset',
-                },
-                [
-                    React.DOM.p(
-                        {
-                            key: 'p1',
-                        },
-                        "We aim to make this site accessible to a wide audience and to ensure a great experience for all users by conforming to World Wide Web Consortium accessibility guidelines (W3C's WCAG)"
-                    ),
-                    React.DOM.p(
-                        {
-                            key: 'p2',
-                        },
-                        'However, if you are having trouble reading this website you can change the way it looks or disable some of its functionalities.'
-                    ),
-                    React.createElement(BinaryToggle, {
-                        key: 'flashing-elements',
-                        name: 'flashing-elements',
-                        label: [
-                            React.DOM.strong(
-                                {
-                                    key: 'label',
-                                },
-                                'Allow flashing elements'
-                            ),
+        return (
+            <form className="form">
+                <fieldset className="fieldset">
+                    <p key="p1">
+                        We aim to make this site accessible to a wide audience
+                        and to ensure a great experience for all users by
+                        conforming to World Wide Web Consortium accessibility
+                        guidelines (W3C&apos;s WCAG)
+                    </p>
+                    <p key="p2">
+                        However, if you are having trouble reading this website
+                        you can change the way it looks or disable some of its
+                        functionalities.
+                    </p>
+                    <BinaryToggle
+                        key="flashing-elements"
+                        name="flashing-elements"
+                        label={[
+                            <strong key="label">
+                                Allow flashing elements
+                            </strong>,
                             this.state['flashing-elements']
                                 ? ' Untick this to disable flashing and moving elements'
                                 : ' Tick this to enable flashing or moving elements.',
-                        ],
-                        enabled: this.state['flashing-elements'],
-                        handleChange: this.toggle.bind(
+                        ]}
+                        enabled={this.state['flashing-elements']}
+                        handleChange={this.toggle.bind(
                             this,
                             'flashing-elements'
-                        ),
-                    }),
-                ]
-            )
+                        )}
+                    />
+                </fieldset>
+            </form>
         );
-    },
-});
+    }
+}
 
 const init = (callback: () => void): void => {
-    React.render(
-        React.createElement(Accessibility),
-        document.getElementById(DOM_ID),
-        callback
-    );
+    const el = document.getElementById(DOM_ID);
+
+    if (el) {
+        render(<Accessibility />, el, callback);
+    }
 };
 
 export { DOM_ID, init };
