@@ -1,14 +1,14 @@
 // @flow
-import { createClass, createElement } from 'react/addons';
+import React, { Component } from 'react';
 import { gridSize } from 'common/modules/crosswords/helpers';
 import { constants } from 'common/modules/crosswords/constants';
 import { classNames } from 'common/modules/crosswords/classNames';
 
-const Cell = createClass({
-    onClick(event) {
+class Cell extends Component<*, *> {
+    onClick(event: Event) {
         event.preventDefault();
         this.props.handleSelect(this.props.x, this.props.y);
-    },
+    }
 
     render() {
         const top = gridSize(this.props.y);
@@ -16,57 +16,54 @@ const Cell = createClass({
 
         let cellNumber = null;
         if (this.props.number !== undefined) {
-            cellNumber = createElement(
-                'text',
-                {
-                    x: left + 1,
-                    y: top + constants.numberSize,
-                    key: 'number',
-                    className: 'crossword__cell-number',
-                },
-                this.props.number
+            cellNumber = (
+                <text
+                    x={left + 1}
+                    y={top + constants.numberSize}
+                    key="number"
+                    className="crossword__cell-number">
+                    {this.props.number}
+                </text>
             );
         }
 
         let cellValue = null;
         if (this.props.value !== undefined) {
-            cellValue = createElement(
-                'text',
-                {
-                    x: left + constants.cellSize * 0.5,
-                    y: top + constants.cellSize * 0.675,
-                    key: 'entry',
-                    className: classNames({
+            cellValue = (
+                <text
+                    x={left + constants.cellSize * 0.5}
+                    y={top + constants.cellSize * 0.675}
+                    key="entry"
+                    className={classNames({
                         'crossword__cell-text': true,
                         'crossword__cell-text--focussed': this.props.isFocussed,
                         'crossword__cell-text--error': this.props.isError,
-                    }),
-                    textAnchor: 'middle',
-                },
-                this.props.value
+                    })}
+                    textAnchor="middle">
+                    {this.props.value}
+                </text>
             );
         }
 
-        return createElement(
-            'g',
-            {
-                onClick: this.onClick,
-            },
-            createElement('rect', {
-                x: left,
-                y: top,
-                width: constants.cellSize,
-                height: constants.cellSize,
-                className: classNames({
-                    crossword__cell: true,
-                    'crossword__cell--focussed': this.props.isFocussed,
-                    'crossword__cell--highlighted': this.props.isHighlighted,
-                }),
-            }),
-            cellNumber,
-            cellValue
+        return (
+            <g onClick={this.onClick.bind(this)}>
+                <rect
+                    x={left}
+                    y={top}
+                    width={constants.cellSize}
+                    height={constants.cellSize}
+                    className={classNames({
+                        crossword__cell: true,
+                        'crossword__cell--focussed': this.props.isFocussed,
+                        'crossword__cell--highlighted': this.props
+                            .isHighlighted,
+                    })}
+                />
+                {cellNumber}
+                {cellValue}
+            </g>
         );
-    },
-});
+    }
+}
 
 export default Cell;
