@@ -1,61 +1,54 @@
 // @flow
-import { createElement, createClass, findDOMNode } from 'react/addons';
+import React, { Component } from 'react';
+import { findDOMNode } from 'react-dom';
 import bean from 'bean';
 import fastdom from 'fastdom';
 import { classNames } from 'common/modules/crosswords/classNames';
 import { isBreakpoint } from 'lib/detect';
 import { scrollTo } from 'lib/scroller';
 
-const Clue = createClass({
+class Clue extends Component<*, *> {
     onClick() {
         this.props.setReturnPosition();
-    },
+    }
 
     render() {
-        return createElement(
-            'li',
-            null,
-            createElement(
-                'a',
-                {
-                    href: `#${this.props.id}`,
-                    onClick: this.onClick,
-                    className: classNames({
+        return (
+            <li>
+                <a
+                    href={`#${this.props.id}`}
+                    onClick={this.onClick.bind(this)}
+                    className={classNames({
                         crossword__clue: true,
                         'crossword__clue--answered': this.props.hasAnswered,
                         'crossword__clue--selected': this.props.isSelected,
                         'crossword__clue--display-group-order':
                             JSON.stringify(this.props.number) !==
                             this.props.humanNumber,
-                    }),
-                },
-                createElement(
-                    'div',
-                    {
-                        className: 'crossword__clue__number',
-                    },
-                    this.props.humanNumber
-                ),
-                createElement('div', {
-                    className: 'crossword__clue__text',
-                    dangerouslySetInnerHTML: {
-                        __html: this.props.clue,
-                    },
-                })
-            )
+                    })}>
+                    <div className="crossword__clue__number">
+                        {this.props.humanNumber}
+                    </div>
+                    <div
+                        className="crossword__clue__text"
+                        dangerouslySetInnerHTML={{ __html: this.props.clue }}
+                    />
+                </a>
+            </li>
         );
-    },
-});
+    }
+}
 
-const Clues = createClass({
-    getInitialState() {
-        return {
+class Clues extends Component<*, *> {
+    constructor(props: Object) {
+        super(props);
+        this.state = {
             showGradient: true,
         };
-    },
+    }
 
     componentDidMount() {
-        this.$cluesNode = findDOMNode(this.refs.clues);
+        this.$cluesNode = (findDOMNode(this.refs.clues): any);
 
         const height =
             this.$cluesNode.scrollHeight - this.$cluesNode.clientHeight;
@@ -69,12 +62,12 @@ const Clues = createClass({
                 });
             }
         });
-    },
+    }
 
     /**
      * Scroll clues into view when they're activated (i.e. clicked in the grid)
      */
-    componentDidUpdate(prev) {
+    componentDidUpdate(prev: Object) {
         if (
             isBreakpoint({
                 min: 'tablet',
@@ -86,11 +79,13 @@ const Clues = createClass({
                 this.scrollIntoView(this.props.focussed);
             });
         }
-    },
+    }
+
+    $cluesNode: HTMLElement;
 
     scrollIntoView(clue: Object) {
         const buffer = 100;
-        const node = findDOMNode(this.refs[clue.id]);
+        const node: HTMLElement = (findDOMNode(this.refs[clue.id]): any);
         const visible =
             node.offsetTop - buffer > this.$cluesNode.scrollTop &&
             node.offsetTop + buffer <
@@ -100,88 +95,52 @@ const Clues = createClass({
             const offset = node.offsetTop - this.$cluesNode.clientHeight / 2;
             scrollTo(offset, 250, 'easeOutQuad', this.$cluesNode);
         }
-    },
+    }
 
     render() {
         const headerClass = 'crossword__clues-header';
         const cluesByDirection = direction =>
             this.props.clues
                 .filter(clue => clue.entry.direction === direction)
-                .map(clue =>
-                    createElement(Clue, {
-                        ref: clue.entry.id,
-                        id: clue.entry.id,
-                        key: clue.entry.id,
-                        number: clue.entry.number,
-                        humanNumber: clue.entry.humanNumber,
-                        clue: clue.entry.clue,
-                        hasAnswered: clue.hasAnswered,
-                        isSelected: clue.isSelected,
-                        setReturnPosition: () => {
+                .map(clue => (
+                    <Clue
+                        ref={clue.entry.id}
+                        id={clue.entry.id}
+                        key={clue.entry.id}
+                        number={clue.entry.number}
+                        humanNumber={clue.entry.humanNumber}
+                        clue={clue.entry.clue}
+                        hasAnswered={clue.hasAnswered}
+                        isSelected={clue.isSelected}
+                        setReturnPosition={() => {
                             this.props.setReturnPosition(window.scrollY);
-                        },
-                    })
-                );
+                        }}
+                    />
+                ));
 
-        return createElement(
-            'div',
-            {
-                className: `crossword__clues--wrapper ${
+        return (
+            <div
+                className={`crossword__clues--wrapper ${
                     this.state.showGradient ? '' : 'hide-gradient'
-                }`,
-            },
-            createElement(
-                'div',
-                {
-                    className: 'crossword__clues',
-                    ref: 'clues',
-                },
-                createElement(
-                    'div',
-                    {
-                        className: 'crossword__clues--across',
-                    },
-                    createElement(
-                        'h3',
-                        {
-                            className: headerClass,
-                        },
-                        'Across'
-                    ),
-                    createElement(
-                        'ol',
-                        {
-                            className: 'crossword__clues-list',
-                        },
-                        cluesByDirection('across')
-                    )
-                ),
-                createElement(
-                    'div',
-                    {
-                        className: 'crossword__clues--down',
-                    },
-                    createElement(
-                        'h3',
-                        {
-                            className: headerClass,
-                        },
-                        'Down'
-                    ),
-                    createElement(
-                        'ol',
-                        {
-                            className: 'crossword__clues-list',
-                        },
-                        cluesByDirection('down')
-                    )
-                )
-            ),
-            createElement('div', {
-                className: 'crossword__clues__gradient',
-            })
+                }`}>
+                <div className="crossword__clues" ref="clues">
+                    <div className="crossword__clues--across">
+                        <h3 className={headerClass}>Across</h3>
+                        <ol className="crossword__clues-list">
+                            {cluesByDirection('across')}
+                        </ol>
+                    </div>
+                    <div className="crossword__clues--down">
+                        <h3 className="headerClass">Down</h3>
+                        <ol className="crossword__clues-list">
+                            {cluesByDirection('down')}
+                        </ol>
+                    </div>
+                </div>
+                <div className="crossword__clues__gradient" />
+            </div>
         );
-    },
-});
+    }
+}
 
 export { Clues };
