@@ -6,6 +6,7 @@ import conf.Configuration
 import model.Cached.{RevalidatableResult, WithoutRevalidationResult}
 import model._
 import com.gu.identity.model.{EmailNewsletter, EmailNewsletters}
+import com.typesafe.scalalogging.LazyLogging
 import play.api.data.Forms._
 import play.api.data._
 import play.api.data.format.Formats._
@@ -13,6 +14,7 @@ import play.api.data.validation.Constraints.emailAddress
 import play.api.libs.json._
 import play.api.libs.ws.{WSClient, WSResponse}
 import play.api.mvc._
+import play.filters.csrf.{CSRFAddToken, CSRFCheck}
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
@@ -30,15 +32,26 @@ case class EmailForm(
   email: String,
   listName: Option[String],
   referrer: Option[String],
-  campaignCode: Option[String])
+  campaignCode: Option[String],
+  name: Option[String]) {
 
-class EmailFormService(wsClient: WSClient) {
+  // `name` is a hidden (via css) form input
+  // if it was set to something this form was likely filled by a bot
+  // https://stackoverflow.com/a/34623588/2823715
+  def isLikelyBotSubmission: Boolean = name.map(_.trim) match {
+    case Some("") | Some(null) | Some("undefined") | None | Some("null") => false
+    case _ => true
+  }
 
-  def submit(form: EmailForm): Future[WSResponse] = {
+}
 
+class EmailFormService(wsClient: WSClient) extends LazyLogging {
+
+  def submit(form: EmailForm): Future[WSResponse] = if (form.isLikelyBotSubmission) {
+    Future.failed(new IllegalAccessException("Form was likely submitted by a bot."))
+  } else {
     val idAccessClientToken = Configuration.id.apiClientToken
     val consentMailerUrl = s"${Configuration.id.apiRoot}/consent-email"
-
     val consentMailerPayload = JsObject(Json.obj("email" -> form.email, "set-lists" -> List(form.listName)).fields)
 
     wsClient
@@ -48,14 +61,16 @@ class EmailFormService(wsClient: WSClient) {
   }
 }
 
-class EmailSignupController(wsClient: WSClient, val controllerComponents: ControllerComponents)(implicit context: ApplicationContext) extends BaseController with ImplicitControllerExecutionContext with Logging {
-    val emailFormService = new EmailFormService(wsClient)
+class EmailSignupController(wsClient: WSClient, val controllerComponents: ControllerComponents, csrfCheck: CSRFCheck, csrfAddToken: CSRFAddToken)(implicit context: ApplicationContext) extends BaseController with ImplicitControllerExecutionContext with Logging {
+  val emailFormService = new EmailFormService(wsClient)
+
   val emailForm: Form[EmailForm] = Form(
     mapping(
       "email" -> nonEmptyText.verifying(emailAddress),
       "listName" -> optional[String](of[String]),
       "referrer" -> optional[String](of[String]),
-      "campaignCode" -> optional[String](of[String])
+      "campaignCode" -> optional[String](of[String]),
+      "name" -> optional(text)
     )(EmailForm.apply)(EmailForm.unapply)
   )
 
@@ -63,25 +78,29 @@ class EmailSignupController(wsClient: WSClient, val controllerComponents: Contro
     Cached(60)(RevalidatableResult.Ok(views.html.emailLanding(emailLandingPage)))
   }
 
-  def renderForm(emailType: String, listId: Int): Action[AnyContent] = Action { implicit request =>
+  def renderForm(emailType: String, listId: Int): Action[AnyContent] = csrfAddToken {
+    Action { implicit request =>
+      val identityName = EmailNewsletter(listId)
+        .orElse(EmailNewsletter.fromV1ListId(listId))
+        .map(_.identityName)
 
-    val identityName = EmailNewsletter(listId)
-                        .orElse(EmailNewsletter.fromV1ListId(listId))
-                        .map(_.identityName)
-
-    identityName match {
-      case Some(listName) => Cached(1.day)(RevalidatableResult.Ok(views.html.emailFragment(emailLandingPage, emailType, listName)))
-      case _ => Cached(15.minute)(WithoutRevalidationResult(NotFound))
+      identityName match {
+        case Some(listName) => Cached(1.day)(RevalidatableResult.Ok(views.html.emailFragment(emailLandingPage, emailType, listName)))
+        case _ => Cached(15.minute)(WithoutRevalidationResult(NotFound))
+      }
     }
   }
 
-  def renderFormFromName(emailType: String, listName: String): Action[AnyContent] = Action { implicit request =>
-    val id = EmailNewsletter.fromIdentityName(listName).map(_.listIdV1)
-    id match {
-      case Some(listId) => Cached(1.day)(RevalidatableResult.Ok(views.html.emailFragment(emailLandingPage, emailType, listName)))
-      case _            => Cached(15.minute)(WithoutRevalidationResult(NotFound))
+  def renderFormFromName(emailType: String, listName: String): Action[AnyContent] = csrfAddToken {
+    Action { implicit request =>
+      val id = EmailNewsletter.fromIdentityName(listName).map(_.listIdV1)
+      id match {
+        case Some(listId) => Cached(1.day)(RevalidatableResult.Ok(views.html.emailFragment(emailLandingPage, emailType, listName)))
+        case _            => Cached(15.minute)(WithoutRevalidationResult(NotFound))
+      }
     }
   }
+
 
   def subscriptionResult(result: String): Action[AnyContent] = Action { implicit request =>
     Cached(7.days)(result match {
@@ -122,21 +141,23 @@ class EmailSignupController(wsClient: WSClient, val controllerComponents: Contro
         Future.successful(respond(InvalidEmail))},
 
       form => emailFormService.submit(form).map(_.status match {
-          case 200 | 201 =>
-            EmailSubmission.increment()
-            respond(Subscribed)
+        case 200 | 201 =>
+          EmailSubmission.increment()
+          respond(Subscribed)
 
-          case status =>
-            log.error(s"Error posting to ExactTarget: HTTP $status")
-            APIHTTPError.increment()
-            respond(OtherError)
+        case status =>
+          log.error(s"Error posting to ExactTarget: HTTP $status")
+          APIHTTPError.increment()
+          respond(OtherError)
 
-        }) recover {
-          case e: Exception =>
-            log.error(s"Error posting to ExactTarget: ${e.getMessage}")
-            APINetworkError.increment()
-            respond(OtherError)
-        })
+      }) recover {
+        case _: IllegalAccessException =>
+          respond(Subscribed)
+        case e: Exception =>
+          log.error(s"Error posting to ExactTarget: ${e.getMessage}")
+          APINetworkError.increment()
+          respond(OtherError)
+      })
   }
 
   def options(): Action[AnyContent] = Action { implicit request =>
