@@ -5,12 +5,7 @@ import { dfpEnv } from 'commercial/modules/dfp/dfp-env';
 import { loadAdvert, refreshAdvert } from 'commercial/modules/dfp/load-advert';
 import { updateAdvertMetric } from 'commercial/modules/dfp/performance-logging';
 import { getAdvertById } from 'commercial/modules/dfp/get-advert-by-id';
-import { getTestVariantId } from 'common/modules/experiments/utils.js';
-import config from 'lib/config';
-import { getViewport, getBreakpoint } from 'lib/detect';
-import fastdom from 'lib/fastdom-promise';
 import { getCurrentTime } from 'lib/user-timing';
-import { session } from 'lib/storage';
 import once from 'lodash/functions/once';
 
 const IntersectionObserver = window.IntersectionObserver;
@@ -26,68 +21,6 @@ const displayAd = (advertId: string): void => {
             loadAdvert(advert);
         }
     }
-};
-
-const optimisedRootMargin = (viewport: {
-    width: number,
-    height: number,
-}): Object => ({
-    front: {
-        mobile: viewport.height * 2,
-        tablet: viewport.height * 2,
-        desktop: viewport.height * 3 / 2,
-        wide: viewport.height * 3 / 2,
-    },
-    'non-front': {
-        mobile: viewport.height * 3 / 2,
-        tablet: viewport.height * 3 / 2,
-        desktop: viewport.height,
-        wide: viewport.height,
-    },
-});
-
-const calculateLazyLoadingDistance = (viewport: {
-    width: number,
-    height: number,
-}) => {
-    const variant = getTestVariantId('CommercialLazyLoadingExtended');
-
-    if (variant === '400px') {
-        return 400;
-    }
-
-    if (variant === '1vh') {
-        return viewport.height;
-    }
-
-    if (variant === '0.5vh') {
-        return viewport.height / 2;
-    }
-
-    if (variant === '0') {
-        return 0;
-    }
-
-    if (variant === 'richard') {
-        const breakpoint = getBreakpoint();
-        const front = config.page.isFront ? 'front' : 'non-front';
-        return optimisedRootMargin(viewport)[front][breakpoint];
-    }
-
-    if (variant === 'super-richard') {
-        if (session.isAvailable()) {
-            const highViewSlots =
-                session.get('gu.commercial.slotVisibility') || 0;
-            const sessionPageViews =
-                session.get('gu.commercial.pageViews') || 0;
-
-            if (highViewSlots / sessionPageViews > 2.5) {
-                return 400;
-            }
-        }
-    }
-
-    return 200;
 };
 
 const onIntersect = (
@@ -109,18 +42,12 @@ const onIntersect = (
     );
 };
 
-const readViewport = (): Promise<{ width: number, height: number }> =>
-    fastdom.read(() => getViewport());
-
 const getObserver = once(() =>
-    readViewport()
-        .then(calculateLazyLoadingDistance)
-        .then(
-            distance =>
-                new window.IntersectionObserver(onIntersect, {
-                    rootMargin: `${distance}px 0px`,
-                })
-        )
+    Promise.resolve(
+        new window.IntersectionObserver(onIntersect, {
+            rootMargin: '200px 0px',
+        })
+    )
 );
 
 export const enableLazyLoad = (advert: Advert): void =>
