@@ -1,16 +1,19 @@
 // @flow
 import {
     getSync as geolocationGetSync,
-    getSupporterPaymentRegion as geolocationGetSupporterPaymentRegion,
+    getSupporterCountryGroup as geolocationGetSupporterPaymentRegion,
 } from 'lib/geolocation';
 
 const geo: string = geolocationGetSync();
 const geoCountryGroup = geolocationGetSupporterPaymentRegion(geo);
 
 const useSupportDomain = (): boolean =>
-    geoCountryGroup === 'GB' ||
-    geoCountryGroup === 'US' ||
-    geoCountryGroup === 'EU';
+    geoCountryGroup === 'GBPCountries' ||
+    geoCountryGroup === 'UnitedStates' ||
+    geoCountryGroup === 'EURCountries' ||
+    geoCountryGroup === 'International' ||
+    geoCountryGroup === 'NZDCountries' ||
+    geoCountryGroup === 'Canada';
 
 const supportBaseURL = useSupportDomain()
     ? 'https://support.theguardian.com/contribute'
