@@ -1,6 +1,7 @@
 // @flow
 
-import React from 'react/addons';
+import React from 'react';
+import { render } from 'react-dom';
 import bean from 'bean';
 import fastdom from 'lib/fastdom-promise';
 import $ from 'lib/$';
@@ -19,10 +20,8 @@ const initCrosswords = (): void => {
                     );
                 }
 
-                const crosswordComponent = React.render(
-                    React.createElement(Crossword, {
-                        data: JSON.parse(data),
-                    }),
+                const crosswordComponent = render(
+                    <Crossword data={JSON.parse(data)} />,
                     element
                 );
 
