@@ -39,6 +39,7 @@ const classes = {
     form: 'js-email-sub__form',
     inlineLabel: 'js-email-sub__inline-label',
     textInput: 'js-email-sub__text-input',
+    dummyInput: 'js-email-sub__text-name',
     listNameHiddenInput: 'js-email-sub__listname-input',
 };
 
@@ -252,19 +253,22 @@ const submitForm = (
 
     return event => {
         const emailAddress = $(`.${classes.textInput}`, $form).val();
+        const csrfToken = $(`input[name=csrfToken]`, $form).val();
+        const dummy = $(`.${classes.dummyInput}`, $form).val();
         const listName = $(`.${classes.listNameHiddenInput}`, $form);
 
         let analyticsInfo;
 
         event.preventDefault();
-
         if (!state.submitting && validate(emailAddress)) {
             const formData = $form.data('formData');
             const data = `email=${encodeURIComponent(
                 emailAddress
-            )}&campaignCode=${formData.campaignCode}&referrer=${
-                formData.referrer
-            }&listName=${listName.val()}`;
+            )}&name=${encodeURIComponent(dummy)}&campaignCode=${
+                formData.campaignCode
+            }&referrer=${formData.referrer}&csrfToken=${encodeURIComponent(
+                csrfToken
+            )}&listName=${listName.val()}`;
 
             analyticsInfo = `rtrt | email form inline | ${
                 analytics.formType
