@@ -19,7 +19,8 @@ const pageHasBlanketBlacklist = (): boolean =>
     // Prevent the blanket emails from ever showing on certain keywords or sections
     keywordExists(['US elections 2016', 'Football']) ||
     config.get('page.section') === 'film' ||
-    config.get('page.seriesId') === 'world/series/guardian-morning-briefing';
+    config.get('page.seriesId') === 'world/series/guardian-morning-briefing' ||
+    config.get('page.pageId') === '/profile/first-dog-on-the-moon'; // temporary to prevent some spammers to annoy us
 
 const userHasRemoved = (id: string, formType: string): boolean => {
     const currentListPrefs = userPrefs.get(`email-sign-up-${formType}`);
