@@ -1,6 +1,10 @@
 // @flow
 import { makeABTest } from 'common/modules/commercial/contributions-utilities';
 import config from 'lib/config';
+import {
+    isPayingMember,
+    isRecentContributor,
+} from 'common/modules/commercial/user-features';
 
 const campaignTag = 'us-news/series/break-the-cycle';
 
@@ -10,6 +14,10 @@ const tagsMatch = () =>
         .split(',')
         .includes(campaignTag);
 
+const isTargetReader = () => isPayingMember() || isRecentContributor();
+
+const isTargetPage = () => tagsMatch() && !config.page.shouldHideReaderRevenue;
+
 export const acquisitionsEpicUSGunCampaign = makeABTest({
     id: 'AcquisitionsUsGunCampaign2017',
     campaignId: 'epic_us_gun_campaign_2017',
@@ -18,15 +26,19 @@ export const acquisitionsEpicUSGunCampaign = makeABTest({
     expiry: '2018-01-04',
 
     author: 'Guy Dawson',
-    description: 'Show a custom Epic for articles with the US gun campaign tag',
+    description:
+        'Show a custom thank you epic for articles with the US gun campaign tag',
     successMeasure: 'AV2.0',
     idealOutcome:
         'The US gun campaign resonates with our readers, and we continue to provide quality reporting on this important issue',
     audienceCriteria: 'All',
     audience: 1,
     audienceOffset: 0,
-    canRun: tagsMatch,
+    showToContributorsAndSupporters: true,
 
+    canRun() {
+        return isTargetReader() && isTargetPage();
+    },
     variants: [
         {
             id: 'control',
@@ -35,7 +47,7 @@ export const acquisitionsEpicUSGunCampaign = makeABTest({
                 isUnlimited: true,
                 usesIframe: true,
                 template: variant =>
-                    `<iframe src="https://interactive.guim.co.uk/embed/2017/11/break-the-cycle/epic.html"
+                    `<iframe src="https://interactive.guim.co.uk/embed/2018/03/break-the-cycle/epic.html"
                         data-component="${variant.options.componentName}"
                         class="acquisitions-epic-iframe"
                         id="${variant.options.iframeId}">
